requirements.txt
baidu-aip==2.0.0.0
Flask==0.12.2
numpy==1.13.3
opencv-python==3.2.0.8
pandas==0.21.0
requests==2.18.4
scipy==1.0.0
six==1.11.0
Werkzeug==0.13
